Details for hidden thread sanitization
