+++ src/components/app/MenuButtons/Publish.css
@@ -12,7 +12,8 @@
 .menuButtonsShareButton,
 .menuButtonsUploadingButton,
 .menuButtonsPermalinkButton,
-.menuButtonsUploadErrorButton {
+.menuButtonsUploadErrorButton,
+.menuButtonsSecondaryShareButton {
   height: 24px;
   margin-bottom: -24px;
 }
@@ -44,9 +45,12 @@
 }
 
 /* The button width needs to be large enough so that the text "Share Without URLs" fits */
+.currentButtonIsPermalinkButton.currentButtonIsSecondaryShareButton
+  .menuButtonsPermalinkButton,
 .menuButtonsUploadingButtonInner,
 .menuButtonsPermalinkButtonButton,
-.menuButtonsUploadErrorButtonButton {
+.menuButtonsUploadErrorButtonButton,
+.menuButtonsSecondaryShareButtonButton {
   width: 115px;
 }
 
@@ -54,8 +58,6 @@
 "Sharing will be enabled once symbolication is complete" */
 .menuButtonsShareButtonButton {
   min-width: 115px;
-  background-color: #0d8730;
-  color: white;
 }
 
 .menuButtonsUploadingButtonProgress {
@@ -69,6 +71,12 @@
 
 .menuButtonsUploadingButtonProgress::-moz-progress-bar {
   background: var(--internal-uploading-progress-fill-color);
+}
+
+.menuButtonsShareButtonButton,
+.menuButtonsSecondaryShareButtonButton {
+  background-color: #0d8730;
+  color: white;
 }
 
 /* CSSTransition */
@@ -108,6 +116,11 @@
 .currentButtonIsUploadingButton .menuButtonsPermalinkButtonButton {
   opacity: 0;
   transform: translateY(100%);
+}
+
+.currentButtonIsPermalinkButton.currentButtonIsSecondaryShareButton
+  .menuButtonsSecondaryShareButton {
+  margin-left: 115px;
 }
 
 .menuButtonsPublishPanel,
@@ -170,7 +183,7 @@
 
 .menuButtonsPublishDataChoicesLabel {
   display: flex;
-  margin: 4px 0;
+  margin: 7px 0;
 }
 
 .menuButtonsPublishDataChoicesLabel.disabled {
@@ -201,7 +214,6 @@
 .menuButtonsPublishButton {
   display: inline-flex;
   min-width: 132px;
-  box-sizing: content-box;
   align-items: center;
   justify-content: center;
   margin: 0 8px;
@@ -209,15 +221,6 @@
   font-size: 13px;
   text-align: center;
   text-decoration: none;
-}
-
-.menuButtonsPublishButtonDisabled {
-  opacity: 0.6;
-}
-
-.menuButtonsPublishButtonDisabled:hover,
-.menuButtonsPublishButtonDisabled:active:hover {
-  background-color: var(--grey-90-a10);
 }
 
 .menuButtonsPublishButtonsSvg {
@@ -240,12 +243,14 @@
   background: url(../../../../res/img/svg/download.svg) center center no-repeat;
 }
 
-.menuButtonsDownloadButton {
-  transition: color 100ms;
+@keyframes menuButtonsDownloadButtonFadeIn {
+  0% { opacity: 0; }
+  100% { opacity: 1; }
 }
 
-.menuButtonsDownloadButtonIsLoading {
-  color: var(--grey-40);
+.menuButtonsDownloadButton {
+  animation-duration: 60ms;
+  animation-name: menuButtonsDownloadButtonFadeIn;
 }
 
 .menuButtonsPublishUpload {
@@ -326,5 +331,18 @@
 
 .menuButtonsPublishError {
   margin: 10px 0;
+  font-size: 13px;
+}
+
+.menuButtonsPublishLabelSubtext {
+  overflow: hidden;
+  max-width: 350px;
+  color: var(--grey-50);
+  font-size: 11px;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.menuButtonsPublishLabelContent {
   font-size: 13px;
 }
