+++ src/components/app/MenuButtons/Publish.js
@@ -12,7 +12,7 @@ import {
   abortUpload,
   resetUploadState,
 } from '../../../actions/publish';
-import { getShouldSanitizeByDefault } from '../../../profile-logic/process-profile';
+import { getShouldSanitizeByDefault } from '../../../profile-logic/sanitization';
 import { getProfile, getProfileRootRange } from '../../../selectors/profile';
 import {
   getCheckedSharingOptions,
@@ -23,6 +23,7 @@ import {
   getUploadProgressString,
   getUploadUrl,
   getUploadError,
+  getNamesOfRemovedThreads,
 } from '../../../selectors/publish';
 import { assertExhaustiveCheck } from '../../../utils/flow';
 
@@ -51,6 +52,7 @@ type StateProps = {|
   +uploadProgress: string,
   +uploadUrl: string,
   +error: mixed,
+  +namesOfRemovedThreads: string[],
 |};
 
 type DispatchProps = {|
@@ -73,7 +75,11 @@ class MenuButtonsPublishImpl extends React.PureComponent<PublishProps> {
     extension: () => this.props.toggleCheckedSharingOptions('extension'),
   };
 
-  _renderCheckbox(slug: $Keys<CheckedSharingOptions>, label: string) {
+  _renderCheckbox(
+    slug: $Keys<CheckedSharingOptions>,
+    label: string,
+    subtext?: string
+  ) {
     const { checkedSharingOptions } = this.props;
     const isDisabled = !checkedSharingOptions.isFiltering;
     const toggle = this._toggles[slug];
@@ -93,9 +99,24 @@ class MenuButtonsPublishImpl extends React.PureComponent<PublishProps> {
           onChange={toggle}
           checked={checkedSharingOptions[slug]}
         />
-        {label}
+        <div className="menuButtonsPublishLabelContent">
+          {label}
+          {subtext ? (
+            <div className="menuButtonsPublishLabelSubtext" title={subtext}>
+              {subtext}
+            </div>
+          ) : null}
+        </div>{' '}
       </label>
     );
+  }
+
+  _renderNamesOfHiddenThreads(): string {
+    const { namesOfRemovedThreads } = this.props;
+    if (namesOfRemovedThreads.length === 0) {
+      return 'No threads are removed.';
+    }
+    return namesOfRemovedThreads.join(', ');
   }
 
   _renderPublishPanel() {
@@ -149,7 +170,11 @@ class MenuButtonsPublishImpl extends React.PureComponent<PublishProps> {
               Filter out potentially identifying information
             </label>
             <div className="menuButtonsPublishDataChoices">
-              {this._renderCheckbox('hiddenThreads', 'Remove hidden threads')}
+              {this._renderCheckbox(
+                'hiddenThreads',
+                'Remove hidden threads',
+                this._renderNamesOfHiddenThreads()
+              )}
               {this._renderCheckbox(
                 'timeRange',
                 'Remove information out of the time range'
@@ -336,6 +361,7 @@ const options: ExplicitConnectOptions<OwnProps, StateProps, DispatchProps> = {
     uploadProgress: getUploadProgressString(state),
     uploadUrl: getUploadUrl(state),
     error: getUploadError(state),
+    namesOfRemovedThreads: getNamesOfRemovedThreads(state),
   }),
   mapDispatchToProps: {
     toggleCheckedSharingOptions,
